Work of Tom Inzina
You don't want to trash your plane ticket - sometimes, you want to take an existing sheet and make quick changes and then take a portion of it into a PPT.

But, when you're done, and you use a shortcut to delete the sheet you better not delete a similar looking sheet. Right?

What this macro (Add-In) does is keep track of sheets. When you open a workbook that doesn't have tracking enabled, it asks if you want tracking. If you say yes, and set the time, then, you will always be asked for confirmation if a sheet is older than the timer limit you set.

CTRL-ALT-D to delete a sheet
CTRL-SHIFT-C to make a copy of a sheet