Work of Eugene Sergiichuk
If you execute the start-of-block macro, then you end up at the first row (that is, go as few rows upwards as you need to) for which the contents of the cell in this column (same column as the active cell) DO NOT MATCH those of the active cell.

For example, if you are in (taking the attached test case) F14 (#10, Swimming) and you run start-of-block, you should end up at F11 (#8, Swimming) NOT F6. NOTE : It is okay to end up in F12 or F7 - as long as you are at the edge of the block. We just want to end up at the edge. That is left up to you. It just needs to behave consistently (same for end of block as well - same definition should apply - that is, the last one that DOES match or the FIRST one that DOES NOT match are both acceptable, but behavior should be predictable)

Similarly for end of block - you would go down as few rows as possible till you get to the first row for which the cell contents of this column (of the active cell) do NOT match those of the active cell. If you start in F8 (#4, Rugby) and run end-of-block, you should end up in F10 (#6, Soccer, NOT F16)

After macro is done, of course, the active cell is the newly found cell. The new active cell will be in the same column of course.