Work of Daniel Jazbec - waveform builder
To draw digital timing diagrams (showing logic levels change) one can actually use Microsoft Excel because it allows you to control the edges of cells.

However, manually drawing the borders to create the waveform is tedious. We want a way to quickly specify the logic level and get the graphical output in the blink of an eye - using VBA.

So the user specifies inputs in a few cells on one row and selects either some cells including these with the inputs or the entire row and runs the macro, and gets the graphical output.