work of Mikhail Musiyachenko upwork
if you can help me troubleshoot runtime dependence on platform, please get in t
-------------------------------
Cell displays "Hello World". When you select the cell and look at what is shown in the Formula Bar, you see ='Another Sheet Name'!$D$33. We now want to be able to click on this cell and go to that reference cell on that other sheet. We want a shortcut that can make update this cell. (NOTE this task only asks for the subroutine, not the shortcut creation).

Spec (note that formatting is important)

Provide two subroutines

1. Operates on selection
2. Operates on entire worksheet

CHECK IF cell is displaying the content of a cell from another sheet. That is, the cell's contents are something like ='Another sheet name'!$D$68 (VBA might have a way that I don't know of to check if this is the case - rather than looking at the contents as a string and analyzing it :)

IF YES, then make the cell a link to that reference cell (the cell will still show the same in the formula field and updating the reference will change what is shown in this cell) so that, when reader clicks on this cell, she is taken to the reference cell.

If the cell has NON-DEFAULT formatting (font, italics, bold, etc), then DON'T update the format in any way (it should LOOK the same. It only becomes a link). Otherwise, it is okay to allow what Excel usually does when you make a link (change color and add underline) - BUT, REMOVE the UNDERLINE.